requirements.txt
django
django-cors-headers
djangorestframework
pillow
dj-database-url
psycopg2-binary
channels
channels_redis
daphne